update filepaths, add index entry to components for ids-splitter
