+++ src/components/ids-splitter/doc/ids-splitter-markup-legend.png
