+++ src/components/ids-splitter/ids-splitter-pane.scss
