+++ src/components/ids-splitter/ids-splitter.scss
