+++ src/components/ids-splitter/index.js
